pyproject.toml
[tool.poetry]
name = "failure_streak_proto"
version = "0.1.0"
description = ""
authors = ["Jeff Zambory <jeff.zambory@mongodb.com>"]
packages = [
    { include = "failurestreaks", from = "src"},
]

[tool.poetry.scripts]
find-streaks = "failurestreaks.find_streaks_cli:main"

[tool.poetry.dependencies]
python = "^3.7"
"evergreen.py" = "^2.2.4"
pymongo = {extras = ["srv"], version = "^3.11.3"}
click = "^7.1.2"
pydantic = "^1.8.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
